Fix d'un overflow

+++ Editeur/Maps/GameMap.vb
@@ -27,17 +27,17 @@ Public Class GameMap
     End Sub
 
     ' - Comptacte les valeurs X et Y du tiles
-    Public Function EncodeXY(ByVal X As Byte, ByVal Y As Byte) As Long
+    Public Function EncodeXY(ByVal X As UShort, ByVal Y As UShort) As UShort
         Return Y * 20 + X
     End Function
 
     ' - Recupère la valeur X du tiles
-    Public Function DecodeX(ByVal code As Long) As Byte
+    Public Function DecodeX(ByVal code As UShort) As UShort
         Return Math.Floor(code - Math.Floor(code / 20) * 20)
     End Function
 
     ' - Recupère la valeur Y du tiles
-    Public Function DecodeY(ByVal code As Long) As Byte
+    Public Function DecodeY(ByVal code As UShort) As UShort
         Return Math.Floor(code / 20)
     End Function
